some improvements

+++ nix/ink-builder.nix
@@ -1,7 +1,7 @@
 {
   base,
+  cargo-contract,
   pkgs,
-  url,
 }: let
   script = pkgs.writeShellScript "builder" ''
     set -e
@@ -10,15 +10,20 @@
       --profile minimal \
       --component rust-src
 
-    CARGO_TARGET_DIR=/root/cargo-contract cargo install cargo-contract \
-     --git https://github.com/paritytech/cargo-contract \
-     --tag v$CARGO_CONTRACT_VERSION
+    if [ "$CARGO_CONTRACT_VERSION" = "${cargo-contract.version}" ]; then
+      mkdir -p /root/.cargo/bin
+      ln -s ${pkgs.lib.getExe cargo-contract.package} /root/.cargo/bin/cargo-contract
+    else
+      CARGO_TARGET_DIR=/root/cargo-contract cargo install cargo-contract \
+        --git https://github.com/paritytech/cargo-contract \
+        --tag v"$CARGO_CONTRACT_VERSION"
 
-    rm -rf /root/cargo-contract
+      rm -rf /root/cargo-contract
+    fi
 
     mkdir source
 
-    curl $SOURCE_CODE_URL \
+    curl "$SOURCE_CODE_URL" \
       -o source.zip
 
     unzip source.zip \
@@ -28,15 +33,15 @@
 
     shopt -s globstar
     for i in **/*.rs; do
-      curl -f ${url}/files/upload/"$BUILD_SESSION_TOKEN" \
+      curl -f "$API_SERVER_URL"/files/upload/"$BUILD_SESSION_TOKEN" \
         -F "$i"="@$i"
     done
 
-    curl -f ${url}/files/seal/"$BUILD_SESSION_TOKEN" \
+    curl -f "$API_SERVER_URL"/files/seal/"$BUILD_SESSION_TOKEN" \
       -X POST
 
     CARGO_TARGET_DIR=/root/artifacts cargo contract build \
-      --release
+      --release &> >(${pkgs.lib.getExe pkgs.ansifilter})
 
     mv /root/artifacts/ink/*.wasm /root/artifacts/ink/main.wasm
     mv /root/artifacts/ink/*.json /root/artifacts/ink/main.json
@@ -67,6 +72,7 @@ in
         "RUST_VERSION"
         "SOURCE_CODE_URL"
         "BUILD_SESSION_TOKEN"
+        "API_SERVER_URL"
       ];
 
       Volumes = {
